iterate: remove Commands
